+++ packages/graphql/src/Builder/Context/HandlerContextBuilderInfo.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types = 1);
 
-namespace LastDragon_ru\LaraASP\GraphQL\Builder\Directives;
+namespace LastDragon_ru\LaraASP\GraphQL\Builder\Context;
 
 use LastDragon_ru\LaraASP\GraphQL\Builder\BuilderInfo;
