+++ packages/graphql/src/Builder/Context/HandlerContextImplicit.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types = 1);
 
-namespace LastDragon_ru\LaraASP\GraphQL\Builder\Directives;
+namespace LastDragon_ru\LaraASP\GraphQL\Builder\Context;
 
 class HandlerContextImplicit {
     public function __construct(
